futurize fixes: lib2to3.fixes.fix_long
cmd:
find . -name '*.py' | xargs futurize -w -n --no-diff -j6 -f lib2to3.fixes.fix_long

+++ rt/pyem/test_typeconverter.py
@@ -140,7 +140,7 @@ class TestTypeConverter(unittest.TestCase):
         fnum = TestUtil.get_debug_float(0)
         TestUtil.to_emobject({"float": fnum})
 
-        lnum = long(num)
+        lnum = int(num)
         TestUtil.to_emobject({"long": lnum})
 
         fl = get_list("float")
